Add files via upload
This is the current set of files I have for color analysis of live video. The primary code used is ColorAnalysisVideo.py, I included a script I had for separate testing named MiscTesting.py. Though that one is not commented and messier, it may provide a reference to my thought processes and what else I tried. This was done in a Visual Studio environment, so you can open the .sln file via Visual Studio if you have it. HaarCascadeObjectDetection.py and ColorAnalysis.py were used as a reference in making this code, while the eye detection cascade is unnecessary, I included it to check the lag response from my computer while running the program (there was notable lag, can be seen in the eyes). This code seems to successfully note the difference in color if I move a colorful object around my face while running, though the overall difference is minor, can be noted while testing. Default color counter is V2, change name in code to test a threshold version.